% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landscape.R
\name{create_landscape}
\alias{create_landscape}
\title{create_landscape}
\usage{
create_landscape(width, height, value)
}
\arguments{
\item{width, height}{dimension of the raster (ncells). Default values:
\code{width} = 126 and \code{height} = 106.}

\item{value}{default value set for the pixels (\code{value} = 0)}
}
\value{
raster Object
}
\description{
Create an empty landscape as raster with a default value
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
