% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disper.R
\name{disper}
\alias{disper}
\title{Seed input by dispersers}
\usage{
disper(x, xr, nf_value, pp_value)
}
\arguments{
\item{x}{A \code{raster} object with the landscape configured}

\item{xr}{A \code{raster} object with richness computed}

\item{nf_value}{The value of "Natural Forests" class within the raster
(default value = 2)}

\item{pp_value}{The value of "pine plantation" class within the raster
(default value = 1)}
}
\description{
Compute the propagule input from each patch to focal pine plantation using
three classes of dispersers and different equations.
}
\details{
It computes the propagule input from each patch to focal pine plantation
using three classes of dispersers and different equations.
\itemize{The quantity and quality of seed dispersion are influenced by:
  \item Seed sources: seed diversity in seed source patch, and patch size
  \item Disperser: percentage of each disperser
  \item Landscape configuration
}
\itemize{Three classes of dispersers were considered:
  \item small birds, \emph{e.g.} European robin (\emph{Erithacus rubecula}),
  Sardinian warbler (\emph{Sylvia melanocephala})
  \item medium birds, \emph{e.g.} Eurasian jay (\emph{Garrulus glandarius})
  \item mammals, \emph{e.g.} Red fox (\emph{Vulpes vulpes})
}

\itemize{For each type of disperser, different dispersion kernels have been
considered.
  \item Small-sized birds rarely exceed 100 m in distance, and approximately
  50\% of the seeds are dispersed in the first 50 m.
  \item Medium-sized birds disperse 50\% of the seeds over a distance of more
  than 100 m. The Eurasyan jay shows a dispersin range between 5 and
  1000 m for Sierra Nevada mountains (SE Spain). The distance at which
  the maximum dispersion occurs depends on the target patch, being
  approximately 400 me when the target patch is a pine plantation.
  \item Mammals disperse in a range from 0 to more than 1500 m, with
  the dispersion peak at 650 - 700 m. More than 50\% of the seeds
  dispersed by mammals are deposited at distances greater than 495 m.
}
}
\references{
Gómez, JM (2003). Spatial patterns in long-distance dispersal of Quercus
ilex acorns by jays in a heterogeneous landscape. Ecography 26: 573 – 584.
\href{https://doi.org/10.1034/j.1600-0587.2003.03586.x}{10.1034/j.1600-0587.2003.03586.x}

González-Varo JP, López-Bao JV, Guitlán J (2013). Functional diversity among
seed dispersal kernels generated by carnivorous mammals. Journal of Animal
Ecology 82(3): 562 - 571.
\href{https://doi.org/10.1111/1365-2656.12024}{10.1111/1365-2656.12024}

Jordano P, García C, Godoy JA, García-Castaño JL (2007). Differential
contribution of frugivores to complex seed dispersal patterns. Proceedings
of the National Academy of Sciences 104: 3278 – 3282.
\href{https://doi.org/10.1073/pnas.0606793104}{10.1073/pnas.0606793104}

Matías L, Zamora R, Mendoza I, Hódar JA. (2010). Seed dispersal patterns by
large frugivorous mammals in a degraded mosaic landscape. Restoration Ecology
18: 619 – 627.
\href{https://doi.org/10.1111/j.1526-100X.2008.00475.x}{10.1111/j.1526-100X.2008.00475.x}

Pons J, Pausas JG (2007). Acorn dispersal estimated by radio-tracking.
Oecologia 153: 903 – 911.
\href{https://doi.org/10.1007/s00442-007-0788-x}{10.1007/s00442-007-0788-x}

Zamora R, Hódar JA, Matías L, Mendoza I. (2010). Positive adjacency effects
mediated by seed disperser birds in pine plantations. Ecological Applications
20: 1053 – 1060.
\href{https://doi.org/10.1890/09-0055.1}{10.1890/09-0055.1}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
