% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disper_time.R
\name{disper_time}
\alias{disper_time}
\title{Propagule input into pine plantation by time}
\usage{
disper_time(
  msb,
  mmb,
  mma,
  x,
  xr,
  pp_value,
  per_sb,
  per_mb,
  per_ma,
  propaguleInputBird,
  propaguleInputMammal,
  time_span
)
}
\arguments{
\item{msb}{A \code{raster} object with small bird contribution on target
pine plantation}

\item{mmb}{A \code{raster} object with medium bird contribution on target
pine plantation}

\item{mma}{A \code{raster} object with mammal contribution on target
pine plantation}

\item{x}{A \code{raster} object with the landscape configured}

\item{xr}{A \code{raster} object with richness computed}

\item{pp_value}{The value of "Pine plantation" class within the raster
(Default value = 1)}

\item{per_sb}{Percentage of small bird dispersers}

\item{per_mb}{Percentage of medium bird dispersers}

\item{per_ma}{Percentage of mammals dispersers}

\item{propaguleInputBird}{Propagule input by birds (seeds number by year and
square meter) \insertCite{@see @Zamora2010}{diveRpine}}

\item{propaguleInputMammal}{Propagule input by mammals (seeds number by year and
square meter) \insertCite{@see @Zamora2010}{diveRpine}}

\item{time_span}{Numbers of year simulated (maximum 50)}
}
\description{
It computes the seeds input into focal pine plantation along the time.
}
\details{
Compute the propagule input from each natural forest patch to pine
plantation target using three types of disperser and different equations.
The seed input into focal pine plantation depends on landscape configuration:
distance to the seed source and quantity and typology of vegetation patches;
and on the composition of the dispersal community. Data about amount of seed
introduced into pine plantation in Sierra Nevada mountain (SE Spain) comes
from \insertCite{@ @Zamora2010}{diveRpine}.
}
\references{
\insertRef{Zamora2010}{diveRpine}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
