% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape.R
\name{plot_landscape}
\alias{plot_landscape}
\title{plot_landscape}
\usage{
plot_landscape(x, ...)
}
\arguments{
\item{x}{A \code{raster} object}

\item{...}{others arguments (from ggplot2)}
}
\value{
ggplot2 Object
}
\description{
Plot the landscape created by the user
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
