% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_propagule.R
\name{plot_propagule}
\alias{plot_propagule}
\title{plot_propagule}
\usage{
plot_propagule(x, ...)
}
\arguments{
\item{x}{A \code{raster} object}

\item{...}{others arguments (from ggplot2)}
}
\value{
ggplot2 Object
}
\description{
Plot the input propagules into the target pine plantation
}
\details{
This auxiliary function plot the input propagules (number of seeds by
pixel and year) into the target pine plantation.
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
