% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_richness.R
\name{plot_richness}
\alias{plot_richness}
\title{plot_richness}
\usage{
plot_richness(x, ...)
}
\arguments{
\item{x}{A \code{raster} object}

\item{...}{others arguments (from ggplot2)}
}
\value{
ggplot2 Object
}
\description{
Plot the richness values for the landscape created by the user
}
\details{
This auxiliary function plots the plant richness values of the
landscape when the landscape is configured (see \emph{Compute Initial Richness}
button in the app), and also the final richness (after the \emph{Dispersers} and
\emph{Simulation} modules were chosen by the user in the app). In both cases
\code{plot_richness} plots the plant richness for each pixel within each of the
landscape categories.
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
