% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryRaster.R
\name{summaryRaster}
\alias{summaryRaster}
\title{summaryRaster}
\usage{
summaryRaster(x)
}
\arguments{
\item{x}{A \code{raster} object
This function computes the mean, min, max and standard
deviation for a specific raster}
}
\value{
a list with mean, min, max, and sd.
}
\description{
Compute custom summary stats for specific raster
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
