% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLandscape.R
\name{createLandscape}
\alias{createLandscape}
\title{Landscape function}
\usage{
createLandscape(r, size_pp, n_nf, size_nf)
}
\arguments{
\item{r}{A \code{raster} object}

\item{size_pp}{The size of pine plantation expressed as pixel numbers
(\code{integer})}

\item{n_nf}{Number of natural forest patchs to create
(\code{integer}, between 1 and 5)}

\item{size_nf}{The size of natural forest patchs (\code{numeric} value,
between 10 and 500) expressed as number of pixels}
}
\value{
A \code{raster} object with the landscape configured
}
\description{
Create a virtual landscape with different patches.
}
\details{
\itemize{This function creates a virtual landscape composed by:
  \item a target pine plantation
  \item natural forests
  \item crops
  \item shrublands
}
The size of the pine plantation and natural forest patchs are specified
by \code{size_pp} and \code{size_nf}.
First the pine plantation patch is set with the dimensions specified by
\code{size_pp}. Then a number of natural forest patchs (\code{n_nf},
between 1 and 5) are created with a determined size (\code{size_nf}).
The crop patches (a random number between 3 and 8) are then established
according to the remaining space. Finally, remain pixels are assigned to
the shrubland category. The size and proportion of the each patch classes
range according to distribution of those ecosystem in Sierra Nevada (SE
Spain) (see Pérez-Luque et al. 2014; 2019)
}
\references{
Pérez-Luque AJ, Bonet FJ, Pérez-Pérez R, Aspizua R, Lorite J, Zamora R
(2014). Sinfonevada: Dataset of floristic diversity in Sierra Nevada forests
(SE Spain). PhytoKeys 35: 1 – 15.
\href{https://doi.org/10.3897/phytokeys.35.6363}{10.3897/phytokeys.35.6363}

Pérez-Luque AJ, Bonet FJ, Zamora R (2019). Map of ecosystems types in Sierra
Nevada mountain (southern Spain).
\url{https://doi.pangaea.de/10.1594/PANGAEA.910176}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
