% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disper.R
\name{disper}
\alias{disper}
\title{Seed input by dispersers}
\usage{
disper(x, xr, nf_value, pp_value)
}
\arguments{
\item{x}{A \code{raster} object with the landscape configured}

\item{xr}{A \code{raster} object with richness computed}

\item{nf_value}{The value of "Natural Forests" class within the raster
(default value = 2)}

\item{pp_value}{The value of "pine plantation" class within the raster
(default value = 1)}
}
\description{
Compute the propagule input from each patch to focal pine plantation using
three classes of dispersers and different equations.
}
\details{
It computes the propagule input from each patch to focal pine plantation
using three classes of dispersers and different equations.
\itemize{The quantity and quality of seed dispersion are influenced by:
  \item Seed sources: seed diversity in seed source patch, and patch size
  \item Disperser: percentage of each disperser
  \item Landscape configuration
}
\itemize{Three classes of dispersers were considered:
  \item small birds, \emph{e.g.} European robin (\emph{Erithacus rubecula}),
  Sardinian warbler (\emph{Sylvia melanocephala})
  \item medium birds, \emph{e.g.} Eurasian jay (\emph{Garrulus glandarius})
  \item mammals, \emph{e.g.} Red fox (\emph{Vulpes vulpes})
}

\itemize{For each type of disperser, different dispersion kernels have been
considered.
  \item Small-sized birds rarely exceed 100 m in distance, and approximately
  50\% of the seeds are dispersed in the first 50 m.
  \item Medium-sized birds disperse 50\% of the seeds over a distance of more
  than 100 m. The Eurasyan jay shows a dispersin range between 5 and
  1000 m for Sierra Nevada mountains (SE Spain). The distance at which
  the maximum dispersion occurs depends on the target patch, being
  approximately 400 me when the target patch is a pine plantation.
  \item Mammals disperse in a range from 0 to more than 1500 m, with
  the dispersion peak at 650 - 700 m. More than 50\% of the seeds
  dispersed by mammals are deposited at distances greater than 495 m.
}
}
\references{
\insertRef{Gomez2003}{diveRpine}

\insertRef{GonzalezVaro2013}{diveRpine}

\insertRef{Jordano2007}{diveRpine}

\insertRef{Matias2010}{diveRpine}

\insertRef{Pons2007}{diveRpine}

\insertRef{Zamora2010}{diveRpine}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
