% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diveRpine-package.R
\docType{package}
\name{diveRpine-package}
\alias{diveRpine}
\alias{diveRpine-package}
\title{diveRpine: Diversification of Pine Plantations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tool for analyze how different forest-management actions modify 
    the biodiversity-related ecosystem services in Mediterranean mountain 
    pine plantationsUse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajpelu/diveRpine}
  \item Report bugs at \url{https://github.com/ajpelu/diveRpine/issues}
}

}
\author{
\strong{Maintainer}: Antonio J. Pérez-Luque \email{ajperez@ugr.es} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID})

Other contributors:
\itemize{
  \item Regino Zamora \email{rzamora@ugr.es} (\href{https://orcid.org/0000-0002-5049-9968}{ORCID}) [contributor]
}

}
\keyword{internal}
