% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRichness.R
\name{initRichness}
\alias{initRichness}
\title{Set values of Initial Richness}
\usage{
initRichness(
  r,
  draster,
  r_range,
  treedensity,
  pastUse,
  elev,
  rad,
  ftreeden,
  fdist,
  fclim,
  w_past,
  w_treeden,
  w_clim,
  w_dist,
  rescale = TRUE
)
}
\arguments{
\item{r}{A \code{raster} object with the configured landscape}

\item{draster}{A \code{raster} object with values of the distance from the
target patch to natural forest patches.}

\item{r_range}{A \verb{data frame} with three columns: \code{value} of land use
(\code{integer}: 0 = "Other", 1 = "Pine plantation", 2 = "Natural Forests",
3 = "Crop"); \code{lowRich} and \code{upRich} (lower and upper value of the
range of Richness).}

\item{treedensity}{density of the pine plantation (\code{integer})}

\item{pastUse}{the past land use of the pine plantation (\code{character}).
One of "Oak", "Shrubland", "Pasture" or "Crop".}

\item{elev, rad}{elevation and annual radiation of the pine plantation}

\item{ftreeden, fdist, fclim}{functions used to computed the effect of tree
density (\code{ftreeden}), and of the distance to natural forest (\code{fdist}), and of
the climate-proxy variables (\code{fclim}) on richness values within pine plantation.
See details.}

\item{w_dist, w_treeden, w_past, w_clim}{weights applied to the functions that
correct the plant richness values according to the distance to seed source,
the tree density, the past-land use of the pine plantation, and also the
climate-proxy variables of the pine plantation.}

\item{rescale}{If "TRUE" the results are rescaled (0 = min and 1 = max)}
}
\value{
A \code{raster} object with values of initial richness for each
pixel.
}
\description{
This function computes the initial values of richness values for each patch.
}
\details{
This function computes the initial plant richness for the land-use
categories (\emph{e.g.} target pine-plantation, surrounding natural
forests, shrubland and crops). For each land-use category, the richness value
of the pixels in each of the patches is randomly calculated from a range
of potential richness values specified by \code{r_range}. In diveRpine, richness
value for each of the patch classes are calculated considering the range
of possible values found on the study area (Sierra Nevada, southern Spain).
Specifically:\tabular{rrr}{
   value \tab lowRich \tab upRich \cr
   No Data \tab 0.00 \tab 0.00 \cr
   Pine Plantations \tab 12.82 \tab 13.34 \cr
   Natural Forests \tab 13.72 \tab 16.11 \cr
   Crops \tab 1.00 \tab 2.00 \cr
}


A richness value is assigned to each pixel. This value will depend on the
pixel category.
\subsection{Pine plantation}{

The richness values of each pixel of the focal (target) pine
patch depends on:
\itemize{
\item Stand Features: tree density, patch size, past land-use, climate-proxy variables
\item Distance to seed source (landscape configuration)
}
For each pixel \emph{j}, the initial richness value (\eqn{R_{init,j}}) is computed
as \deqn{Richess \sim Potential\ Richenss \times fc}
where \emph{Potential Richenss} is a random value coming from \code{r_range} and \emph{fc}
is a correction factor:
\deqn{fc = w_{past}\cdot f(\textrm{past~Land~Use}) + w_{dist}\cdot f(\textrm{Seed~source~distance}) + w_{treeden}\cdot f(\textrm{Tree~density}) + w_{clim}\cdot f(\textrm{Climate~proxy})}

We specified the following weights according to literature (see references):
\itemize{
\item \eqn{w_{past}} = 0.2
\item \eqn{w_{dist}} = 0.35
\item \eqn{w_{treeden}} = 0.25
\item \eqn{w_{clim}} = 0.2
}

but different weights can be provided using \code{w_past}, \code{w_dist},\code{w_treeden},
and \code{w_clim} respectively.

Each of the factors affecting the richness within a pine plantation pixel are
computed as follows:
\subsection{Tree density (\code{ftreeden})}{

Richness and species diversity within pine plantation are strongly conditioned
by tree density, which has a negative effect on the plant diversity, and on
the total plant species richness. Potential richness is affected as a function
of density, as follows:
\deqn{\textrm{ftreeden} = \exp \left(-\frac{1}{2} \left( \frac{ \textrm{treeDensity} - 0.22} {1504.1}\right )^2\right )}

This equation is the used by default, but could be change with \code{ftreeden}. Tree
density value is specified by \code{treedensity}.
}

\subsection{Climate-proxy factors}{

Potential richness are also strongly affected by climate-proxy variables.
It has been determined from 19 climatic and topographical variables
that the elevation and the annual radiation are the variables that best
explaining the variability of potential richness within pine-plantation
(they capturing more than 83.3 \% of the observed variance). The climate effect
on potential richness within pine-plantation has been modeled according to the
following equation:
\deqn{\textrm{fclim}=\textrm{exp}\left (-\frac{1}{2}\left ( \frac{\textrm{Altitude}- 1557.16}{644.89} \right )^{2}  \right ) \times ~\textrm{exp}\left (-\frac{1}{2}\left ( \frac{\textrm{Radiation}}{13.24} \right )^{2}  \right )}

This equation is the used by default, but could be change with \code{fclim}. Altitude
and radiation values are specified by \code{elev} and \code{rad}.
}

\subsection{Seed source distance (\code{fdist}).}{

Seed dispersal depends on the distance from the seed source. In pine plantations,
the presence and abundance of species other than pines is determined, among
others, by the distance to the seed source. In Sierra Nevada (southern Spain)
natural oak forests are the most influential in terms of distance to the seed
source. Oak vegetation has higher plant diversity than pine plantations,
especially for herbaceous species. Shorter distances could increase the pool
of species in the pine plantations and reduce the evenness of plantation
communities. The relationship found between distance to the source and diversity
observed in pine plantations is governed by the following equation:
\deqn{\textrm{Diversity} = 1.7605 - 0.0932 * \sqrt{\sqrt{\textrm{Distance}}}}

This equation is the used by default, but could be change with \code{fdist}. For
each pixel of pine plantation the distances between the centroid of the pixel
and the edge of each natural forest patches are computed using the function
\code{\link[=dist2nf]{dist2nf()}} which generate a distance raster (\code{draster}).
}

\subsection{Past Land Use}{

The richness value of a plantation is conditioned by the past land use. For
instance, it has been found that regeneration of \emph{Quercus} in pine
plantations depends more on past land-use than on plantation tree density and
distance to the seed source. Navarro-González et al. (2013), found that the
probability finding regeneration within a plantation varies as a function of
past land use. We rescaled the gradient found by Navarro-González et al. (2013)
as follow:  natural forest (0.9999), Shrubland (0.4982), Cropland (0.0279),
and Grassland (0.0001). The value of \eqn{f(\textrm{past Land Use})} is selected
according to the past land use specified in \code{pastUse}.
}

}

\subsection{Natural forests and Croplands}{

The initial richness values of each pixel of natural forest and cropland patches are
randomly selected from the value ranges specified on \code{r_range}.
}
}
\references{
\insertRef{GomezAparicio2009}{diveRpine}

\insertRef{Mendoza2009}{diveRpine}

\insertRef{GonzalezMoreno2011}{diveRpine}

\insertRef{NavarroGonzalez2013}{diveRpine}

\insertRef{PerezLuque2014}{diveRpine}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
