% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRichness.R
\name{initRichness}
\alias{initRichness}
\title{Set values of Initial Richness}
\usage{
initRichness(r, draster, r_range, treedensity, pastUse, rescale = TRUE)
}
\arguments{
\item{r}{A \code{raster} object with the configured landscape}

\item{draster}{A \code{raster} object with values of the distance from the
target patch to natural forest patchs}

\item{r_range}{A \code{data frame} with three columns: \code{value} of land use
(\code{integer}: 0 = "Other", 1 = "Pine plantation", 2 = "Natural Forests",
3 = "Crop"); \code{lowRich} and \code{upRich} (lower and upper value of the
range of Richness: See Gómez-Aparicio et al. 2009)}

\item{treedensity}{density of the pine plantation (\code{integer})}

\item{pastUse}{the past land use of the pine plantation (\code{character}).
One of "Oak", "Shrubland", "Pasture" or "Crop"}

\item{rescale}{If "TRUE" the results are rescaled (0 = min and 1 = max)}
}
\value{
A \code{raster} object with values of initial richness for each
pixel.
}
\description{
This function computes the initial values of richness values for each patch.
}
\details{
This function computes the initial richness for each land-use
categories. The richenss values of each pixel of the focal (target) pine
patch depends on:
\itemize{
  \item Stand structure: tree density, patch size, past land-use
  \item Distance to seed source (landscape configuration)
}

Richenss value for each of the patch classes (\emph{i.e.} pine plantation,
natural forests, shrubland and crops) are calculated considering the range
of possible values found on the study area. In this case we use data from
Sierra Nevada (southern Spain) See References.
}
\references{
Gómez-Aparicio L, Zavala MA, Bonet FJ, Zamora R (2009). Are pine plantations
valid tools for restoring Mediterranean forests? An assessment along abiotic
and biotic gradients. Ecological Applications, 19: 2124 - 2141.
\href{https://doi.org/10.1890/08-1656.1}{10.1890/08-1656.1}

Mendoza I, Gómez-Aparicio L, Zamora R, Matías L (2009). Recruitment
limitation of forest communities in a degraded Mediterranean landscape.
Journal of Vegetation Science 20: 367 – 376.
\href{https://doi.org/10.1111/j.1654-1103.2009.05705.x}{10.1111/j.1654-1103.2009.05705.x}

González-Moreno P, Quero JL, Poorter L, Bonet FJ, Zamora R (2011). Is spatial
structure the key to promote plant diversity in Mediterranean forest
plantations? Basic and Applied Ecology 12: 251 – 259.
\href{https://doi.org/10.1016/j.baae.2011.02.012}{10.1016/j.baae.2011.02.012}

Navarro-González I, Pérez-Luque AJ, Bonet FJ, Zamora R (2013). The weight of
the past: land-use legacies and recolonization of pine plantations by oak
trees. Ecological Applications 23: 1267 – 1276.
\href{https://doi.org/10.1890/12-0459.1}{10.1890/12-0459.1}

Pérez-Luque AJ, Bonet FJ, Pérez-Pérez R, Aspizua R, Lorite J, Zamora R
(2014). Sinfonevada: Dataset of floristic diversity in Sierra Nevada forests
(SE Spain). PhytoKeys 35: 1 – 15.
\href{https://doi.org/10.3897/phytokeys.35.6363}{10.3897/phytokeys.35.6363}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
