% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_dispersion.R
\name{potential_dispersion}
\alias{potential_dispersion}
\title{Compute the potential dispersion}
\usage{
potential_dispersion(
  x,
  nf_value,
  pp_value,
  rich_nf,
  kernel_sbi,
  kernel_mbi,
  kernel_ma,
  seedlim_int,
  seedlim_slope
)
}
\arguments{
\item{x}{A \code{raster} object with the landscape configured}

\item{nf_value}{The value of "Natural Forests" class within the raster
(default value = 2).}

\item{pp_value}{The value of "pine plantation" class within the raster
(default value = 1).}

\item{rich_nf}{A \code{raster} object with richness of the natural forests}

\item{kernel_sbi, kernel_mbi, kernel_ma}{The dispersion kernel functions for
small birds (\code{kernel_sbi}), medium birds (\code{kernel_mbi}) and mammals (\code{kernel_ma}).
See details.}

\item{seedlim_int, seedlim_slope}{The intercept and slope of the linear relation
between the seed limitation in the target plantation and the adjacency of a
natural forest patch with the target pine plantation. See details}
}
\value{
\code{raster} objects with values of potential dispersion for each
type of disperser
}
\description{
Compute the propagule input from each patch to target pine plantation using
three classes of disperses and different kernels equations.
}
\details{
It computes the propagule input from each patch to focal pine plantation
using three classes of disperses and different kernels.
\itemize{The quantity and quality of seed dispersion are influenced by:
\item Seed sources: seed diversity in seed source patch, and patch size
\item Disperses: percentage of each disperser type
\item Landscape configuration
}

\itemize{Three classes of disperses were considered by default for the study
area:
\item small birds, \emph{e.g.} European robin (\emph{Erithacus rubecula}),
Sardinian warbler (\emph{Sylvia melanocephala})
\item medium birds, \emph{e.g.} Eurasian jay (\emph{Garrulus glandarius})
\item mammals, \emph{e.g.} Red fox (\emph{Vulpes vulpes})
}

\itemize{For each type of disperser, different dispersion kernels have been
considered:
\item Small-sized birds rarely exceed 100 m in distance, and approximately
50\% of the seeds are dispersed in the first 50 m.
\item Medium-sized birds disperse 50\% of the seeds over a distance of more
than 100 m. The Eurasyan jay shows a dispersion range between 5 and 1000 m
for Sierra Nevada mountains (SE Spain). The distance at which the maximum
dispersion occurs depends on the target patch, being approximately 400 m
when the target patch is a pine plantation.
\item Mammals disperse in a range from 0 to more than 1500 m, with
the dispersion peak at 650 - 700 m. More than 50\% of the seeds
dispersed by mammals are deposited at distances greater than 495 m.
}

According to the disperser type, the function uses by default different
dispersion kernels. For small and medium birds a log-normal dispersion kernels
were used. Specifically, log-normal density function with mean and standard
deviation values of log(51) and log(2) respectively for small-sized birds,
and log-normal density function with mean and standard deviation values of
log(201) and log(2) for medium-sized birds. This distribution were implemented
using the \code{\link[stats:Lognormal]{stats::dlnorm()}} function. For mammals, a combination of log-normal
and Weibull dispersion kernel was used. Specifically, from 0 to 400 m of distance
from the seed source, a Weibull distribution with a shape and scale parameters
of 1.385 and 137 respectively was used (See \code{\link[stats:Weibull]{stats::dweibull()}}); whereas a
log-normal density function with mean and standard deviation values of log(751)
and log(1.346) respectively, were applied for distances higher than 401 m
from seed source. Different kernels functions could be specified using the
parameters \code{kernel_sbi}, \code{kernel_mbi}, and \code{kernel_ma}.

This function also considers the adjacency between each of the natural forest
patches and the target pine plantation. The higher the adjacency between the
natural forest and the pine plantation, the lower the limitation of the propagule
entry dispersed by birds. Zamora et al. (2010) found that the intercept and the
slope of the linear relation between the seed limitation (values from 0 to 1)
and the adjacency (\emph{i.e.} percentage of pine-plantation perimeter in contact
with native forests) were 0.733 and 0.0039 respectively. Those parameters are
included by default, but could be customized using the \code{seedlim_int} and
\code{seedlim_slope} parameters. For each natural forest patch, the adjacency to
target pine-plantation is computed. For those patches with adjacency, the
potential dispersion by birds increases according a correction factor (see
Zamora et al. 2010), that is computed as follows:

\eqn{adj_{fc} = 1 + \frac{\textrm{seed Entry} - \textrm{seed Entry}_{0}}{\textrm{seed Entry}_{100} - \textrm{seed Entry}_{0}}}

where \eqn{\textrm{seed Entry} = 1- \textrm{seed limitation}}; \eqn{\textrm{seed limitation}}
is computed using the \code{seedlim_int} and \code{seedlim_slope} parameters;
\eqn{\textrm{seed Entry}_{0}} and \eqn{\textrm{seed Entry}_{100}} correspond
to the seed entry for no adjacency and full adjacency respectively (see
Zamora et al. 2010).
}
\references{
\insertRef{Gomez2003}{diveRpine}

\insertRef{GonzalezVaro2013}{diveRpine}

\insertRef{Jordano2007}{diveRpine}

\insertRef{Matias2010}{diveRpine}

\insertRef{Pons2007}{diveRpine}

\insertRef{Zamora2010}{diveRpine}
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
